brio-package {brio}	R Documentation

brio: Basic R Input Output

Description

Functions to handle basic input output, these functions always read and write UTF-8 (8-bit Unicode Transformation Format) files and provide more explicit control over line endings.

Author(s)

Maintainer: Gábor Csárdi csardi.gabor@gmail.com

Authors:

Jim Hester (ORCID)

Other contributors:

Posit Software, PBC [copyright holder, funder]

See Also

Useful links:

https://brio.r-lib.org

https://github.com/r-lib/brio

Report bugs at https://github.com/r-lib/brio/issues

[Package brio version 1.1.4 Index]