Import

Read file by extension into R.

import(file, ...)

Arguments

file
character(1). File path.

...
Additional arguments, passed to the respective internal loading function.

Current recommendations (by priority):

data.frame: Base R default. Generally recommended.

S3 class.

Allows rownames, but they're required and can't be set NULL.

See help(topic = "data.frame", package = "base") for details.

DataFrame: Recommended when working with Bioconductor packages.

S4 class; inherits DataTable virtual class.

Allows rownames, but they're optional and can be set NULL.

See help(topic = "DataFrame", package = "S4Vectors") for details.

tbl_df (tibble): Recommended when working with tidyverse packages.

S3 class; inherits data.frame.

Does not allow rownames.

See help(topic = "tibble", package = "tibble") for details.

data.table: Recommended when working with the data.table package.

S3 class; inherits data.frame.

Does not allow rownames.

See help(topic = "data.table", package = "data.table") for details.

Value

Varies, depending on the file extension:

Default: DataFrame.

MTX: sparseMatrix.

GTF/GFF: GRanges.

JSON/YAML: list.

Source code or log: character.

Details

Remote URLs and compressed files are supported. All extensions are case insensitive.

This function supports automatic loading of common file types:

CSV: Comma Separated Values.
Imported by data.table::fread().

TSV Tab Separated Values.
Imported by data.table::fread().

TXT: Text file. Ambiguous, not recommended.
Imported by data.table::fread().

XLSX/XLS: Excel workbook.
Imported by readxl::read_excel().

MTX: MatrixMarket sparse matrix.
Imported by Matrix::readMM().

GTF/GFF/GFF3: General Feature Format.
Imported by rtracklayer::import().

JSON: JSON. Imported by jsonlite::read_json().

YAML/YML: YAML. Imported by yaml::yaml.load_file().

RDA/RDATA: R Data.

Imported by load.

Must contain a single object.

Doesn't require internal object name to match, unlike loadData().

RDS: R Data Serialized.
Imported by readRDS().

These file formats will be imported as source code lines by readr::read_lines(): LOG, MD, PY, R, RMD, SH.

These file formats are blacklisted, and intentionally not supported: DOC, DOCX, PDF, PPT, PPTX.

If a file format isn't supported natively (or blacklisted), the rio package will be used as a fallback attempt. See rio::import() for details.

Delimited Files (CSV/TSV)

import() uses the fread() function of the data.table package to import standard CSV and TSV files. This should work automatically for most files without issue.

Here are some notable exceptions:

Columns headers should be character. In the event that a column name is numeric, set header = TRUE here to force the column name.

See help(topic = "fread", package = "data.table") for details.

The read_csv() and read_tsv() functions of the readr package are good alternatives, which return tibble data frames (tbl_df).

Google Sheets

Public Google Sheets are now supported. Simply paste in the URL.

Matrix Market Exchange (MTX/MEX)

Reading a Matrix Market Exchange (MTX) file now requires COLNAMES and ROWNAMES sidecar files containing the colnames and rownames of the sparse matrix. Legacy support for manual loading of these sidecar files is provided.

General Feature Format (GFF/GTF)

The GFF (General Feature Format) format consists of one line per feature, each containing 9 columns of data, plus optional track definition lines. The GTF (General Transfer Format) is identical to GFF version 2.

Column names follow the Ensembl conventions. Additional information:

Ensembl

Gencode

GMT/GMX files

Refer to the Broad Institute GSEA wiki for details.

bcbio files

Also supports some additional extensions commonly used with the bcbio pipeline:

COUNTS: Counts table (e.g. RNA-seq aligned counts).

COLNAMES: Sidecar file containing column names.

ROWNAMES: Sidecar file containing row names.

Data frame return

By default, import() returns a standard data.frame. However, any of these desired output formats can be set globally using brio.data.frame:

data.frame

DataFrame (S4Vectors).

data.table (data.table).

tbl_df (tibble).

See also

Examples

file <- system.file("extdata/example.csv", package = "brio")
x <- import(file)

#> Importing example.csv using data.table::fread().

#> Setting rownames from `rowname` column.

print(x)

#>        sample1 sample2 sample3 sample4
#> gene1       16      20      13      16
#> gene2       29      22      43      50
#> gene3      243     245     186     184
#> gene4        7      14      25      16
#> gene5        1       1       2       2
#> gene6       10      17      18      11
#> gene7        1       2       0       3
#> gene8        8      27       7       4
#> gene9        1       0       6       2
#> gene10      13       0       1       2
#> gene11      30      12       6       5
#> gene12      25      25      14      34
#> gene13      13      57      21      25
#> gene14     162      85      85      62
#> gene15       0       2       9       1
#> gene16      28      13      26      12
#> gene17     142     392     376     215
#> gene18     147      92      97     129
#> gene19       5       4       7       0
#> gene20       3       0       9       0
#> gene21      16       4       8       7
#> gene22      34      22      47      44
#> gene23       1       0       0       0
#> gene24      18      46      39      49
#> gene25       3      19      16      25
#> gene26      39      93      35      88
#> gene27      17      26      18      39
#> gene28      48     147      68     117
#> gene29       4       3       1       0
#> gene30      20      18      21       8
#> gene31       7       1      11       9
#> gene32     131     114      59     133
#> gene33      12      27      13      13
#> gene34      20      10      10      15
#> gene35      47      56      22      52
#> gene36      14      48      54      76
#> gene37      26       3      46      10
#> gene38      10      16      34      18
#> gene39       0       0       0       4
#> gene40       6       9       0       9
#> gene41       0       1       1       2
#> gene42       3       2      11       6
#> gene43      16       4      14      23
#> gene44       1       6       7       5
#> gene45      33      26       8       9
#> gene46       8       6       0      19
#> gene47      37      70      51      10
#> gene48      47      20      14      53
#> gene49      26      69      52      40
#> gene50       1       0       0       2