Export

Export data out of R and write to disk.

# S4 method for data.frame
export(x, file, format)

# S4 method for DataFrame
export(x, file, format)

# S4 method for matrix
export(x, file, format)

# S4 method for sparseMatrix
export(x, file, format)

# S4 method for matrix
export(x, file, format)

# S4 method for SummarizedExperiment
export(x, name = NULL, dir = ".",
  compress = FALSE, slotNames = c("assays", "colData", "rowData"))

# S4 method for SingleCellExperiment
export(x, name = NULL, dir = ".",
  compress = FALSE, slotNames = c("assays", "colData", "rowData"))

Arguments

x
Object. An object supporting dim(), or a supported class capable of being coerced to data.frame, to be written to disk.

file
character(1). File path. Specify file or format but not both.

format
character(1). An optional character string containing the file format, which can be used to override the format inferred from file, or in lieu of specifying file.

name
character(1). Name to use on disk. If NULL, will use the name of the object instead.

dir
character(1). Directory path.

compress
logical(1). Apply gzip compression to all files.

slotNames
character. Names of slots to include when writing to disk.

Value

Invisible character. File path(s).

Row names

Some export utilities in R have a tendency to drop row names when writing to disk in CSV format. For example, the readr family of functions never write row names by design. This is a really poor default setting for handling genomic data, which often contain gene identifiers in the row names. Here we're performing any internal tibble coercion step to ensure row names are always moved to a "rowname" column in the CSV export.

Exporting vectors

Use writeLines() instead of export() to write vectors to disk. An S4 character method may be defined in a future update, but it is intentionally unsupported in the current release.

See also

Examples

counts <- matrix(data = seq_len(100L), nrow = 10)
export(counts, format = "csv")

#> Exported counts.csv.

## Clean up.
file.remove("counts.csv")

#> [1] TRUE